# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: APIs-Graduation
  version: 1.0.0
paths:
...
  /api/medical-test/medical-tests:
    get:
      tags:
        - 'Medical Tests'
      summary: 'Get all medical tests'
      operationId: 3327fd32528d4dcdd3346c93d54c7bb1
      responses:
        200:
          description: 'Medical tests retrieved successfully'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      properties:
                        id:
                          type: integer
                          example: 1
                        test_name:
                          type: string
                          example: 'Blood Test'
                        description:
                          type: string
                          example: 'A test to analyze blood components'
                          nullable: true
                        schedule:
                          oneOf:
                            -
                              $ref: '#/components/schemas/DoctorSchedule'
                          nullable: true
                        cost:
                          type: number
                          format: float
                          example: 50
                      type: object
                  message:
                    type: string
                    example: 'Medical tests retrieved successfully.'
                  status:
                    type: integer
                    example: 200
                type: object
        401:
          description: Unauthenticated
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Unauthenticated
                  status:
                    type: integer
                    example: 401
                type: object
        403:
          description: 'Unauthorized action'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Unauthorized. Only doctors or patients can perform this action.'
                  status:
                    type: integer
                    example: 403
                type: object
      security:
        -
          bearerAuth: []
  /api/medical-test/create/medical-test:
    post:
      tags:
        - 'Medical Tests'
      summary: 'Create a new medical test'
      operationId: 6258682d755b482d79d0f27640b223d9
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - test_name
                - cost
              properties:
                test_name:
                  description: 'Name of the medical test'
                  type: string
                  example: 'Blood Test'
                description:
                  description: 'Description of the medical test'
                  type: string
                  example: 'A test to analyze blood components'
                  nullable: true
                schedule:
                  oneOf:
                    -
                      $ref: '#/components/schemas/DoctorSchedule'
                  nullable: true
                  description: 'Schedule in JSON format'
                cost:
                  description: 'Cost of the medical test'
                  type: number
                  format: float
                  example: 50
              type: object
      responses:
        201:
          description: 'Medical test created successfully'
          content:
            application/json:
              schema:
                properties:
                  data:
                    properties:
                      id:
                        type: integer
                        example: 1
                      test_name:
                        type: string
                        example: 'Blood Test'
                      description:
                        type: string
                        example: 'A test to analyze blood components'
                        nullable: true
                      schedule:
                        oneOf:
                          -
                            $ref: '#/components/schemas/DoctorSchedule'
                        nullable: true
                      cost:
                        type: number
                        format: float
                        example: 50
                    type: object
                  message:
                    type: string
                    example: 'Medical test created successfully.'
                  status:
                    type: integer
                    example: 201
                type: object
        401:
          description: Unauthenticated
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Unauthenticated
                  status:
                    type: integer
                    example: 401
                type: object
        422:
          description: 'Validation error'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Validation error.'
                  errors:
                    type: object
                  status:
                    type: integer
                    example: 422
                type: object
      security:
        -
          bearerAuth: []
  '/api/medical-test/medical-test/view/{id}':
    get:
      tags:
        - 'Medical Tests'
      summary: 'Get a specific medical test'
      operationId: 897942a24504bbae8c792e911dfb0f3a
      parameters:
        -
          name: id
          in: path
          description: 'ID of the medical test'
          required: true
          schema:
            type: string
      responses:
        200:
          description: 'Medical test retrieved successfully'
          content:
            application/json:
              schema:
                properties:
                  data:
                    properties:
                      id:
                        type: integer
                        example: 1
                      test_name:
                        type: string
                        example: 'Blood Test'
                      description:
                        type: string
                        example: 'A test to analyze blood components'
                        nullable: true
                      schedule:
                        oneOf:
                          -
                            $ref: '#/components/schemas/DoctorSchedule'
                        nullable: true
                      cost:
                        type: number
                        format: float
                        example: 50
                    type: object
                  message:
                    type: string
                    example: 'Medical test retrieved successfully.'
                  status:
                    type: integer
                    example: 200
                type: object
        401:
          description: Unauthenticated
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Unauthenticated
                  status:
                    type: integer
                    example: 401
                type: object
        403:
          description: 'Unauthorized action'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Unauthorized. Only patients can perform this action.'
                  status:
                    type: integer
                    example: 403
                type: object
        404:
          description: 'Medical test not found'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Medical test not found.'
                  status:
                    type: integer
                    example: 404
                type: object
      security:
        -
          bearerAuth: []
  '/api/medical-test/medical-test/update/{id}':
    put:
      tags:
        - 'Medical Tests'
      summary: 'Update a specific medical test'
      operationId: c30cb7022f77bd807592cc8a76306c76
      parameters:
        -
          name: id
          in: path
          description: 'ID of the medical test'
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                test_name:
                  description: 'Name of the medical test'
                  type: string
                  example: 'Blood Test'
                description:
                  description: 'Description of the medical test'
                  type: string
                  example: 'A test to analyze blood components'
                  nullable: true
                schedule:
                  oneOf:
                    -
                      $ref: '#/components/schemas/DoctorSchedule'
                  nullable: true
                  description: 'Schedule in JSON format'
                cost:
                  description: 'Cost of the medical test'
                  type: number
                  format: float
                  example: 50
              type: object
      responses:
        200:
          description: 'Medical test updated successfully'
          content:
            application/json:
              schema:
                properties:
                  data:
                    properties:
                      id:
                        type: integer
                        example: 1
                      test_name:
                        type: string
                        example: 'Blood Test'
                      description:
                        type: string
                        example: 'A test to analyze blood components'
                        nullable: true
                      schedule:
                        oneOf:
                          -
                            $ref: '#/components/schemas/DoctorSchedule'
                        nullable: true
                      cost:
                        type: number
                        format: float
                        example: 50
                    type: object
                  message:
                    type: string
                    example: 'Medical test updated successfully.'
                  status:
                    type: integer
                    example: 200
                type: object
        401:
          description: Unauthenticated
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Unauthenticated
                  status:
                    type: integer
                    example: 401
                type: object
        404:
          description: 'Medical test not found'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Medical test not found.'
                  status:
                    type: integer
                    example: 404
                type: object
        422:
          description: 'Validation error'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Validation error.'
                  errors:
                    type: object
                  status:
                    type: integer
                    example: 422
                type: object
      security:
        -
          bearerAuth: []
  '/api/medical-test/medical-test/delete/{id}':
    delete:
      tags:
        - 'Medical Tests'
      summary: 'Delete a specific medical test'
      operationId: 26404e528673316c911d7c5ec5c86d3a
      parameters:
        -
          name: id
          in: path
          description: 'ID of the medical test'
          required: true
          schema:
            type: string
      responses:
        200:
          description: 'Medical test deleted successfully'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Medical test deleted successfully.'
                  status:
                    type: integer
                    example: 200
                type: object
        401:
          description: Unauthenticated
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Unauthenticated
                  status:
                    type: integer
                    example: 401
                type: object
        404:
          description: 'Medical test not found'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Medical test not found.'
                  status:
                    type: integer
                    example: 404
                type: object
      security:
        -
          bearerAuth: []
...
components:
  schemas:
    DoctorSchedule:
      properties:
        Monday:
          type: array
          items:
            type: string
            example: '10:00 AM, 1:00PM, 4:00PM'
        Tuesday:
          type: array
          items:
            type: string
            example: '9:00 AM, 1:00PM, 4:00PM'
      type: object
...